Extract e2e assertion helpers
