+++ frontend/test/__support__/e2e/cypress.js
@@ -15,6 +15,7 @@ export * from "./helpers/e2e-qa-databases-helpers";
 export * from "./helpers/e2e-ad-hoc-question-helpers";
 export * from "./helpers/e2e-enterprise-helpers";
 export * from "./helpers/e2e-mock-app-settings-helpers";
+export * from "./helpers/e2e-assertion-helpers";
 
 export function setupLocalHostEmail() {
   // Email info
@@ -91,16 +92,6 @@ export function setupDummySMTP() {
     "email-smtp-username": "nevermind",
     "email-smtp-password": "it-is-secret-NOT",
     "email-from-address": "nonexisting@metabase.test",
-  });
-}
-
-export function expectedRouteCalls({ route_alias, calls } = {}) {
-  const requestsCount = alias =>
-    cy.state("requests").filter(req => req.alias === alias);
-  // It is hard and unreliable to assert that something didn't happen in Cypress
-  // This solution was the only one that worked out of all others proposed in this SO topic: https://stackoverflow.com/a/59302542/8815185
-  cy.get("@" + route_alias).then(() => {
-    expect(requestsCount(route_alias)).to.have.length(calls);
   });
 }
